fix(ObjectPage): correct position of snapped `image`
