+++ packages/main/src/components/ObjectPage/ObjectPage.module.css
@@ -47,11 +47,7 @@
   z-index: 4;
   cursor: pointer;
   display: grid;
-
-  [data-component-name='ObjectPageTitle'] {
-    grid-column: 2;
-    padding-inline: 0;
-  }
+  grid-auto-columns: 100%;
 }
 
 /*:has cannot be nested*/
@@ -185,7 +181,6 @@
 }
 
 .snappedContent {
-  grid-column: 1 / span 2;
   padding-block-end: 0.5rem;
 }
