+++ packages/main/src/components/ObjectPage/index.tsx
@@ -590,27 +590,23 @@ const ObjectPage = forwardRef<ObjectPageDomRef, ObjectPagePropTypes>((props, ref
         data-not-clickable={!!preserveHeaderStateOnClick}
         aria-roledescription={accessibilityAttributes?.objectPageTopHeader?.ariaRoledescription ?? 'Object Page header'}
         className={classNames.header}
-        style={{
-          gridAutoColumns: `min-content ${
-            titleArea && image && headerCollapsed === true ? `calc(100% - 3rem - 1rem)` : '100%'
-          }`
-        }}
       >
         <span
           className={classNames.clickArea}
           onClick={onTitleClick}
           data-component-name="ObjectPageTitleAreaClickElement"
         />
-        {titleArea && image && headerCollapsed === true && (
-          <CollapsedAvatar image={image} imageShapeCircle={imageShapeCircle} />
-        )}
         {titleArea &&
           cloneElement(titleArea as ReactElement<ObjectPageTitlePropsWithDataAttributes>, {
             className: clsx(titleArea?.props?.className),
             onToggleHeaderContentVisibility: onTitleClick,
             'data-not-clickable': !!preserveHeaderStateOnClick,
             'data-header-content-visible': headerArea && headerCollapsed !== true,
-            'data-is-snapped-rendered-outside': snappedHeaderInObjPage
+            'data-is-snapped-rendered-outside': snappedHeaderInObjPage,
+            _snappedAvatar:
+              titleArea && image && headerCollapsed === true ? (
+                <CollapsedAvatar image={image} imageShapeCircle={imageShapeCircle} />
+              ) : null
           })}
         {snappedHeaderInObjPage && (
           <div className={classNames.snappedContent} data-component-name="ATwithImageSnappedContentContainer">
